fully working version
- "faces" record when it hits the left edge of the mine in the supply
curve as well
features left to implement
- Program the plotting and saving of the simulation results and the
optimal policy/value results, so that results can be easily viewed and
saved.

- Program works as expected with simplified supply curve. Can explain
the results.
○ Does the best policy make sense? Can I explain it by stepping through
the different steps manually and arrive at the result?
○ Is program symmetrical? (flip the ownership assignment in the supply
curves and see if we get the same answer) (flip the order of doing
things if the mines are the same and see if we get the same results)